Modularize dir-merge parser into focused submodules
