--- crates/engine/src/local_copy/dir_merge/parse.rs
@@ -1,602 +1,0 @@
-use crate::local_copy::filter_program::{
-    DirMergeEnforcedKind, DirMergeOptions, ExcludeIfPresentRule,
-};
-use rsync_filters::FilterRule;
-use std::error::Error;
-use std::fmt;
-use std::path::PathBuf;
-
-#[derive(Debug)]
-pub(crate) enum ParsedFilterDirective {
-    Rule(FilterRule),
-    Merge {
-        path: PathBuf,
-        options: Option<DirMergeOptions>,
-    },
-    ExcludeIfPresent(ExcludeIfPresentRule),
-    Clear,
-}
-
-#[derive(Debug)]
-pub(crate) struct FilterParseError {
-    message: String,
-}
-
-impl FilterParseError {
-    pub(crate) fn new(message: impl Into<String>) -> Self {
-        Self {
-            message: message.into(),
-        }
-    }
-}
-
-impl fmt::Display for FilterParseError {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        f.write_str(&self.message)
-    }
-}
-
-impl Error for FilterParseError {}
-
-pub(crate) fn parse_filter_directive_line(
-    text: &str,
-) -> Result<Option<ParsedFilterDirective>, FilterParseError> {
-    if text.is_empty() || text.starts_with('#') {
-        return Ok(None);
-    }
-
-    let trimmed = text.trim_start();
-    if trimmed.is_empty() {
-        return Ok(None);
-    }
-
-    let trimmed = trimmed.trim_end();
-
-    if trimmed == "!" || trimmed.eq_ignore_ascii_case("clear") {
-        return Ok(Some(ParsedFilterDirective::Clear));
-    }
-
-    if let Some(directive) = parse_short_merge_directive_line(trimmed)? {
-        return Ok(Some(directive));
-    }
-
-    if let Some(directive) = parse_merge_directive(trimmed)? {
-        return Ok(Some(directive));
-    }
-
-    if let Some(directive) = parse_dir_merge_directive(trimmed)? {
-        return Ok(Some(directive));
-    }
-
-    const EXCLUDE_IF_PRESENT_PREFIX: &str = "exclude-if-present";
-
-    if trimmed.len() >= EXCLUDE_IF_PRESENT_PREFIX.len()
-        && trimmed[..EXCLUDE_IF_PRESENT_PREFIX.len()]
-            .eq_ignore_ascii_case(EXCLUDE_IF_PRESENT_PREFIX)
-    {
-        let mut remainder = trimmed[EXCLUDE_IF_PRESENT_PREFIX.len()..]
-            .trim_start_matches(|ch: char| ch == '_' || ch.is_ascii_whitespace());
-        if let Some(rest) = remainder.strip_prefix('=') {
-            remainder = rest.trim_start_matches(|ch: char| ch == '_' || ch.is_ascii_whitespace());
-        }
-
-        let pattern_text = remainder.trim();
-        if pattern_text.is_empty() {
-            return Err(FilterParseError::new(
-                "filter directive 'exclude-if-present' requires a marker file",
-            ));
-        }
-
-        return Ok(Some(ParsedFilterDirective::ExcludeIfPresent(
-            ExcludeIfPresentRule::new(pattern_text),
-        )));
-    }
-
-    if let Some(remainder) = trimmed.strip_prefix('+') {
-        let pattern = remainder.trim_start();
-        if pattern.is_empty() {
-            return Err(FilterParseError::new("filter rule '+' requires a pattern"));
-        }
-        return Ok(Some(ParsedFilterDirective::Rule(FilterRule::include(
-            pattern.to_string(),
-        ))));
-    }
-
-    if let Some(remainder) = trimmed.strip_prefix('-') {
-        let pattern = remainder.trim_start();
-        if pattern.is_empty() {
-            return Err(FilterParseError::new("filter rule '-' requires a pattern"));
-        }
-        return Ok(Some(ParsedFilterDirective::Rule(FilterRule::exclude(
-            pattern.to_string(),
-        ))));
-    }
-
-    let mut parts = trimmed.splitn(2, char::is_whitespace);
-    let keyword = parts.next().unwrap_or("");
-    let remainder = parts.next().unwrap_or("").trim_start();
-
-    let handle_keyword = |pattern: &str,
-                          builder: fn(String) -> FilterRule|
-     -> Result<Option<ParsedFilterDirective>, FilterParseError> {
-        if pattern.is_empty() {
-            return Err(FilterParseError::new("filter directive missing pattern"));
-        }
-        Ok(Some(ParsedFilterDirective::Rule(builder(
-            pattern.to_string(),
-        ))))
-    };
-
-    if keyword.len() == 1 {
-        let shorthand = keyword.chars().next().unwrap().to_ascii_lowercase();
-        match shorthand {
-            'p' => {
-                return handle_keyword(remainder, FilterRule::protect);
-            }
-            'r' => {
-                return handle_keyword(remainder, FilterRule::risk);
-            }
-            's' => {
-                if remainder.is_empty() {
-                    return Err(FilterParseError::new("filter directive missing pattern"));
-                }
-                let rule = FilterRule::show(remainder.to_string());
-                return Ok(Some(ParsedFilterDirective::Rule(rule)));
-            }
-            'h' => {
-                if remainder.is_empty() {
-                    return Err(FilterParseError::new("filter directive missing pattern"));
-                }
-                let rule = FilterRule::hide(remainder.to_string());
-                return Ok(Some(ParsedFilterDirective::Rule(rule)));
-            }
-            _ => {}
-        }
-    }
-
-    if keyword.eq_ignore_ascii_case("include") {
-        return handle_keyword(remainder, FilterRule::include);
-    }
-
-    if keyword.eq_ignore_ascii_case("exclude") {
-        return handle_keyword(remainder, FilterRule::exclude);
-    }
-
-    if keyword.eq_ignore_ascii_case("show") {
-        if remainder.is_empty() {
-            return Err(FilterParseError::new("filter directive missing pattern"));
-        }
-        let rule = FilterRule::show(remainder.to_string());
-        return Ok(Some(ParsedFilterDirective::Rule(rule)));
-    }
-
-    if keyword.eq_ignore_ascii_case("hide") {
-        if remainder.is_empty() {
-            return Err(FilterParseError::new("filter directive missing pattern"));
-        }
-        let rule = FilterRule::hide(remainder.to_string());
-        return Ok(Some(ParsedFilterDirective::Rule(rule)));
-    }
-
-    if keyword.eq_ignore_ascii_case("protect") {
-        return handle_keyword(remainder, FilterRule::protect);
-    }
-
-    if keyword.eq_ignore_ascii_case("risk") {
-        return handle_keyword(remainder, FilterRule::risk);
-    }
-
-    Err(FilterParseError::new(format!(
-        "unsupported filter directive '{}'",
-        trimmed
-    )))
-}
-
-pub(crate) fn parse_merge_directive(
-    text: &str,
-) -> Result<Option<ParsedFilterDirective>, FilterParseError> {
-    const MERGE_PREFIX: &str = "merge";
-
-    if text.len() < MERGE_PREFIX.len() {
-        return Ok(None);
-    }
-
-    let (prefix, rest) = text.split_at(MERGE_PREFIX.len());
-    if !prefix.eq_ignore_ascii_case(MERGE_PREFIX) {
-        return Ok(None);
-    }
-
-    let mut remainder = rest.trim_start_matches(|ch: char| ch == '_' || ch.is_ascii_whitespace());
-    let mut modifiers = "";
-    if let Some(next) = remainder.strip_prefix(',') {
-        let mut split = next.splitn(2, |ch: char| ch.is_ascii_whitespace() || ch == '_');
-        modifiers = split.next().unwrap_or("");
-        remainder = split
-            .next()
-            .unwrap_or("")
-            .trim_start_matches(|ch: char| ch == '_' || ch.is_ascii_whitespace());
-    }
-
-    let (options, assume_cvsignore) = parse_merge_modifiers(modifiers, text, false)?;
-
-    if remainder == "-" {
-        return Err(FilterParseError::new(
-            "merge from standard input is not supported in .rsync-filter files",
-        ));
-    }
-
-    let path_text = remainder.trim_end();
-    let path_text = if path_text.is_empty() {
-        if assume_cvsignore {
-            ".cvsignore"
-        } else {
-            return Err(FilterParseError::new(
-                "merge directive requires a file path",
-            ));
-        }
-    } else {
-        path_text
-    };
-
-    let options = if modifiers.is_empty() && !assume_cvsignore {
-        None
-    } else {
-        Some(options)
-    };
-
-    Ok(Some(ParsedFilterDirective::Merge {
-        path: PathBuf::from(path_text),
-        options,
-    }))
-}
-
-pub(crate) fn parse_short_merge_directive_line(
-    text: &str,
-) -> Result<Option<ParsedFilterDirective>, FilterParseError> {
-    let mut chars = text.chars();
-    let first = match chars.next() {
-        Some(first) => first,
-        None => return Ok(None),
-    };
-
-    let allow_extended = match first {
-        '.' => false,
-        ':' => true,
-        _ => return Ok(None),
-    };
-
-    let remainder = chars.as_str();
-    let (modifiers, rest) = split_short_rule_modifiers(remainder);
-    let (options, assume_cvsignore) = parse_merge_modifiers(modifiers, text, allow_extended)?;
-
-    let pattern = rest.trim();
-    let pattern = if pattern.is_empty() {
-        if assume_cvsignore {
-            ".cvsignore"
-        } else if allow_extended {
-            return Err(FilterParseError::new(format!(
-                "dir-merge directive '{}' is missing a file name",
-                text
-            )));
-        } else {
-            return Err(FilterParseError::new(format!(
-                "merge directive '{}' is missing a file path",
-                text
-            )));
-        }
-    } else {
-        pattern
-    };
-
-    if allow_extended {
-        return Ok(Some(ParsedFilterDirective::Merge {
-            path: PathBuf::from(pattern),
-            options: Some(options),
-        }));
-    }
-
-    let options = if modifiers.is_empty() && !assume_cvsignore {
-        None
-    } else {
-        Some(options)
-    };
-
-    Ok(Some(ParsedFilterDirective::Merge {
-        path: PathBuf::from(pattern),
-        options,
-    }))
-}
-
-pub(crate) fn split_short_rule_modifiers(text: &str) -> (&str, &str) {
-    if text.is_empty() {
-        return ("", "");
-    }
-
-    if let Some(rest) = text.strip_prefix(',') {
-        let mut parts = rest.splitn(2, |ch: char| ch.is_ascii_whitespace() || ch == '_');
-        let modifiers = parts.next().unwrap_or("");
-        let remainder = parts.next().unwrap_or("");
-        let remainder =
-            remainder.trim_start_matches(|ch: char| ch == '_' || ch.is_ascii_whitespace());
-        return (modifiers, remainder);
-    }
-
-    let mut chars = text.chars();
-    match chars.next() {
-        None => ("", ""),
-        Some(first) if first.is_ascii_whitespace() || first == '_' => {
-            let remainder =
-                text.trim_start_matches(|ch: char| ch == '_' || ch.is_ascii_whitespace());
-            ("", remainder)
-        }
-        Some(_) => {
-            let mut len = 0;
-            for ch in text.chars() {
-                if ch.is_ascii_whitespace() || ch == '_' {
-                    break;
-                }
-                len += ch.len_utf8();
-            }
-            let modifiers = &text[..len];
-            let remainder =
-                text[len..].trim_start_matches(|ch: char| ch == '_' || ch.is_ascii_whitespace());
-            (modifiers, remainder)
-        }
-    }
-}
-
-pub(crate) fn parse_merge_modifiers(
-    modifiers: &str,
-    directive: &str,
-    allow_extended: bool,
-) -> Result<(DirMergeOptions, bool), FilterParseError> {
-    let label = if allow_extended { "dir-merge" } else { "merge" };
-    let mut options = if allow_extended {
-        DirMergeOptions::default()
-    } else {
-        DirMergeOptions::default().allow_list_clearing(true)
-    };
-    let mut enforced: Option<DirMergeEnforcedKind> = None;
-    let mut saw_include = false;
-    let mut saw_exclude = false;
-    let mut assume_cvsignore = false;
-
-    for modifier in modifiers.chars() {
-        let lower = modifier.to_ascii_lowercase();
-        match lower {
-            '-' => {
-                if saw_include {
-                    let message = format!(
-                        "{label} directive '{}' cannot combine '+' and '-' modifiers",
-                        directive
-                    );
-
-                    return Err(FilterParseError::new(message));
-                }
-                saw_exclude = true;
-                enforced = Some(DirMergeEnforcedKind::Exclude);
-            }
-            '+' => {
-                if saw_exclude {
-                    let message = format!(
-                        "{label} directive '{}' cannot combine '+' and '-' modifiers",
-                        directive
-                    );
-                    return Err(FilterParseError::new(message));
-                }
-                saw_include = true;
-                enforced = Some(DirMergeEnforcedKind::Include);
-            }
-            'c' => {
-                if saw_include {
-                    let message = format!(
-                        "{label} directive '{}' cannot combine 'C' with '+' or '-'",
-                        directive
-                    );
-                    return Err(FilterParseError::new(message));
-                }
-                saw_exclude = true;
-                enforced = Some(DirMergeEnforcedKind::Exclude);
-                options = options
-                    .use_whitespace()
-                    .allow_comments(false)
-                    .allow_list_clearing(true)
-                    .inherit(false);
-                assume_cvsignore = true;
-            }
-            'e' => {
-                if allow_extended {
-                    options = options.exclude_filter_file(true);
-                } else {
-                    let message = format!(
-                        "merge directive '{}' uses unsupported modifier '{}'",
-                        directive, modifier
-                    );
-                    return Err(FilterParseError::new(message));
-                }
-            }
-            'n' => {
-                if allow_extended {
-                    options = options.inherit(false);
-                } else {
-                    let message = format!(
-                        "merge directive '{}' uses unsupported modifier '{}'",
-                        directive, modifier
-                    );
-                    return Err(FilterParseError::new(message));
-                }
-            }
-            'w' => {
-                if allow_extended {
-                    options = options.use_whitespace().allow_comments(false);
-                } else {
-                    let message = format!(
-                        "merge directive '{}' uses unsupported modifier '{}'",
-                        directive, modifier
-                    );
-                    return Err(FilterParseError::new(message));
-                }
-            }
-            's' => {
-                if allow_extended {
-                    options = options.sender_modifier();
-                } else {
-                    let message = format!(
-                        "merge directive '{}' uses unsupported modifier '{}'",
-                        directive, modifier
-                    );
-                    return Err(FilterParseError::new(message));
-                }
-            }
-            'r' => {
-                if allow_extended {
-                    options = options.receiver_modifier();
-                } else {
-                    let message = format!(
-                        "merge directive '{}' uses unsupported modifier '{}'",
-                        directive, modifier
-                    );
-                    return Err(FilterParseError::new(message));
-                }
-            }
-            '/' => {
-                if allow_extended {
-                    options = options.anchor_root(true);
-                } else {
-                    let message = format!(
-                        "merge directive '{}' uses unsupported modifier '{}'",
-                        directive, modifier
-                    );
-                    return Err(FilterParseError::new(message));
-                }
-            }
-            _ => {
-                let message = format!(
-                    "{label} directive '{}' uses unsupported modifier '{}'",
-                    directive, modifier
-                );
-                return Err(FilterParseError::new(message));
-            }
-        }
-    }
-
-    options = options.with_enforced_kind(enforced);
-    if !allow_extended && !options.list_clear_allowed() {
-        options = options.allow_list_clearing(true);
-    }
-
-    Ok((options, assume_cvsignore))
-}
-
-pub(crate) fn parse_dir_merge_directive(
-    text: &str,
-) -> Result<Option<ParsedFilterDirective>, FilterParseError> {
-    const DIR_MERGE_PREFIX: &str = "dir-merge";
-
-    if text.len() < DIR_MERGE_PREFIX.len() {
-        return Ok(None);
-    }
-
-    let (prefix, mut remainder) = text.split_at(DIR_MERGE_PREFIX.len());
-    if !prefix.eq_ignore_ascii_case(DIR_MERGE_PREFIX) {
-        return Ok(None);
-    }
-
-    if let Some(ch) = remainder.chars().next() {
-        if ch != ',' && !ch.is_ascii_whitespace() {
-            return Ok(None);
-        }
-    }
-
-    remainder = remainder.trim_start();
-
-    let mut modifiers = "";
-    if let Some(rest) = remainder.strip_prefix(',') {
-        let mut split = rest.splitn(2, char::is_whitespace);
-        modifiers = split.next().unwrap_or("");
-        remainder = split.next().unwrap_or("").trim_start();
-    }
-
-    let mut options = DirMergeOptions::default();
-    let mut saw_plus = false;
-    let mut saw_minus = false;
-    let mut used_cvs_default = false;
-
-    for modifier in modifiers.chars() {
-        let lower = modifier.to_ascii_lowercase();
-        match lower {
-            '-' => {
-                if saw_plus {
-                    let message = format!(
-                        "dir-merge directive '{}' cannot combine '+' and '-' modifiers",
-                        text
-                    );
-                    return Err(FilterParseError::new(message));
-                }
-                saw_minus = true;
-                options = options.with_enforced_kind(Some(DirMergeEnforcedKind::Exclude));
-            }
-            '+' => {
-                if saw_minus {
-                    let message = format!(
-                        "dir-merge directive '{}' cannot combine '+' and '-' modifiers",
-                        text
-                    );
-                    return Err(FilterParseError::new(message));
-                }
-                saw_plus = true;
-                options = options.with_enforced_kind(Some(DirMergeEnforcedKind::Include));
-            }
-            'n' => {
-                options = options.inherit(false);
-            }
-            'e' => {
-                options = options.exclude_filter_file(true);
-            }
-            'w' => {
-                options = options.use_whitespace();
-                options = options.allow_comments(false);
-            }
-            's' => {
-                options = options.sender_modifier();
-            }
-            'r' => {
-                options = options.receiver_modifier();
-            }
-            '/' => {
-                options = options.anchor_root(true);
-            }
-            'c' => {
-                used_cvs_default = true;
-                options = options.with_enforced_kind(Some(DirMergeEnforcedKind::Exclude));
-                options = options.use_whitespace();
-                options = options.allow_comments(false);
-                options = options.inherit(false);
-                options = options.allow_list_clearing(true);
-            }
-            _ => {
-                let message = format!(
-                    "dir-merge directive '{}' uses unsupported modifier '{}'",
-                    text, modifier
-                );
-                return Err(FilterParseError::new(message));
-            }
-        }
-    }
-
-    let path_text = if remainder.is_empty() {
-        if used_cvs_default {
-            ".cvsignore"
-        } else {
-            let message = format!("dir-merge directive '{}' is missing a file name", text);
-            return Err(FilterParseError::new(message));
-        }
-    } else {
-        remainder
-    };
-
-    Ok(Some(ParsedFilterDirective::Merge {
-        path: PathBuf::from(path_text),
-        options: Some(options),
-    }))
-}
