+++ crates/engine/src/local_copy/dir_merge/parse/mod.rs
@@ -1,0 +1,8 @@
+mod dir_merge;
+mod line;
+mod merge;
+mod modifiers;
+mod types;
+
+pub(crate) use line::parse_filter_directive_line;
+pub(crate) use types::{FilterParseError, ParsedFilterDirective};
